+++ src/shared/index.ts
@@ -9,5 +9,6 @@ export * from "./types/execution.js";
 export * from "./types/agent.js";
 export * from "./types/task.js";
 export * from "./types/config.js";
+export * from "./types/connector.js";
 export * from "./types/workflow.js";
 export * from "./constants.js";